refactor(browser): serve snapshot and scroll from the browser crate
Page snapshotting and scrolling ran from vmux_desktop only because the
request types lived in vmux_agent and the mechanism lived in
vmux_browser, and nothing could see both. The mechanism is browser work:
it drives CEF Browsers, resolves panes to webviews through
vmux_layout::target, and maintains PendingNavSnapshots, which was already
in vmux_browser.

Move the four request and response types to vmux_core::browser, which
both crates already depend on and which already owns
dom_snapshot::RawSnapshot. The systems then move to vmux_browser with no
new crate edge, and BrowserAutomationPlugin disappears.

Screenshots and recording merge into MediaPlugin. Both capture the app
for the agent and both carry a compiled-out arm that rejects requests
rather than dropping them.
