+++ crates/vmux_agent/src/events.rs
@@ -1,6 +1,5 @@
 use bevy::prelude::*;
 use serde_json::Value;
-use std::collections::HashSet;
 
 pub use vmux_service::agent_events::{
     AgentCommandRequest, AgentQueryRequest, AgentToolCallRequest, CommandOrigin,
@@ -79,34 +78,6 @@ pub struct ScreenshotResponse {
     pub request_id: [u8; 16],
     pub result: Result<ScreenshotImage, String>,
 }
-
-#[derive(Message, Clone)]
-pub struct BrowserSnapshotRequest {
-    pub request_id: [u8; 16],
-    pub pane: Option<String>,
-    pub webview: Option<Entity>,
-}
-
-#[derive(Message, Clone)]
-pub struct BrowserSnapshotResponse {
-    pub request_id: [u8; 16],
-    pub result: Result<String, String>,
-}
-
-#[derive(Message, Clone)]
-pub struct BrowserScrollRequest {
-    pub request_id: [u8; 16],
-    pub pane: Option<String>,
-    pub to: Option<String>,
-    pub delta: Option<i32>,
-}
-
-/// Request ids whose `BrowserSnapshotResponse` must be returned as an agent
-/// *command* result (a navigation that returns its page snapshot inline)
-/// rather than the default *query* result. Populated when a deferred
-/// navigation settles; drained by `forward_snapshot_responses`.
-#[derive(Resource, Default)]
-pub struct NavAwaitingSnapshot(pub HashSet<[u8; 16]>);
 
 pub fn snapshot_response_to_query_result(
     result: &Result<String, String>,
